requirements.txt
playwright~=1.51.0
Faker~=37.1.0
pytest~=8.3.5
python-dotenv~=1.1.0
requests~=2.32.3
allure-python-commons~=2.14.0